// cerbtest-app/src/app/models/order.ts
export class order {
    orderID? : number;
    customerName = "";
    employeeName = "";
    orderDate?: Date;
    requiredDate?: Date;
    shippedDate?: Date;
    shipVia = "";
    freight? : number;
    shipName = "";
    shipAddress = "";
    shipCity = ""; 
    shipRegion = ""; 
    shipPostalCode = "";
    shipCountry = "";
}